backend/requirements.txt
flask
flask-cors
torch
torchvision
numpy
pandas
scikit-learn
scikit-fuzzy
joblib
pillow
opencv-python
supabase
python-dotenv
google-generativeai